feat: :sparkles: Add rate limiter

+++ app/Http/Controllers/Api/AuthController.php
@@ -10,6 +10,7 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Validator;
 use Nette\Utils\Json;
+use Illuminate\Support\Facades\RateLimiter;
 
 class AuthController extends Controller
 {
@@ -94,27 +95,38 @@ class AuthController extends Controller
     
         $credentials = request(['email', 'password']);
 
-        if (Auth::attempt($credentials)) {
+        try {
 
-            $user = $request->user();
+            if (RateLimiter::tooManyAttempts(request()->ip(), 3)) {
+                return response()->json(
+                    [ 'message' => 'Too Many Failed Login Attempt. Restricted for 30 Seconds'], 
+                );
+            }
 
-            $user->tokens()->delete();
-            
-            $responseData = [
-                'status' => 'success',
-                'message' => 'Successful Login',
-                'data' => [
-                    'token' => $user->createToken(Auth::user())->plainTextToken,
-                    'user' => $user
-                ]
-            ];
+            if (Auth::attempt($credentials)) {
 
-            // print_r (Json::decode($user->tokens[0]->name)->role_id);
+                $user = $request->user();
+    
+                $user->tokens()->delete();
+                
+                $responseData = [
+                    'status' => 'success',
+                    'message' => 'Successful Login',
+                    'data' => [
+                        'token' => $user->createToken(Auth::user())->plainTextToken,
+                        'user' => $user
+                    ]
+                ];
+                RateLimiter::clear(request()->ip());
+                return response($responseData, 200);
+            }
 
-            return response($responseData, 200);
+            RateLimiter::hit(request()->ip(), 30);
+            return response($responseData, 400);
         }
-
-        return response($responseData, 400);
+        catch(\Throwable $th) {
+            throw $th;
+        }
     }
 
     public function logout(Request $request)
